Add initial implementation of workers code on top of postgresql queing.
It is an initial impl, and need to be rethinked in order to
archive better efficiency. The current approach implies have an active
thread for each worker that is not bad as is but is bad when
the workers will be not very active and allocating thread just for
sleep it not make sense.